some renamings and doc

+++ xcmailr-webapp/src/main/java/models/MBox.java
@@ -682,28 +682,36 @@ public class MBox
     {
         String[] boxArray = boxIds.split("\\,");
         StringBuilder query = new StringBuilder();
-        
-        query.append("SELECT m.id, m.address, m.domain FROM MAILBOXES m WHERE ");
-        query.append("m.usr_id = ").append(userId);
-        query.append(" AND (");
-        for (String bId : boxArray)
+        if (boxArray.length > 0)
         {
-            query.append(" id = ").append(bId);
-            query.append(" OR");
+            query.append("SELECT m.id, m.address, m.domain FROM MAILBOXES m WHERE ");
+            query.append("m.usr_id = ").append(userId);
+            query.append(" AND (");
+            for (String bId : boxArray)
+            {
+                query.append(" id = ").append(bId);
+                query.append(" OR");
+            }
+            query.delete(query.length() - 2, query.length());
+            query.append(");");
+
+            RawSql rawSql = RawSqlBuilder.parse(query.toString()).columnMapping("m.id", "id")
+                                         .columnMapping("m.address", "address").columnMapping("m.domain", "domain")
+                                         .create();
+            Query<MBox> quer = Ebean.find(MBox.class).setRawSql(rawSql);
+            List<MBox> selectedBoxes = quer.findList();
+            StringBuilder csvMail = new StringBuilder();
+
+            for (MBox mailBox : selectedBoxes)
+            {
+                csvMail.append(mailBox.getFullAddress()).append("\n");
+            }
+            return csvMail.toString();
         }
-        query.delete(query.length() - 2, query.length());
-        query.append(");");
-        
-        RawSql rawSql = RawSqlBuilder.parse(query.toString()).columnMapping("m.id", "id").columnMapping("m.address", "address").columnMapping("m.domain", "domain").create();
-        Query<MBox> quer = Ebean.find(MBox.class).setRawSql(rawSql);
-        List<MBox> selectedBoxes = quer.findList();
-        StringBuilder csvMail = new StringBuilder();
-    
-        for (MBox mailBox : selectedBoxes)
+        else
         {
-            csvMail.append(mailBox.getFullAddress()).append("\n");
+            return "";
         }
-        return csvMail.toString();
     }
 
     /**
